+++ UI/Pane.h
@@ -8,13 +8,13 @@
 
 class Pane : public Widget {
 public:
-	Pane(double x, double y, std::string title);
+	Pane(float x, float y, std::string title);
 
 	void add(Widget *widget);
 
 	void draw() override;
 
-	bool contains(int x, int y) override;
+	bool contains(float x, float y) override;
 
 	bool notify(Event *e) override;
 
@@ -22,7 +22,7 @@ private:
 	std::string title;
 	std::vector<Widget *> widgets;
 	bool selected;
-	int marginLeft, marginTop, dx, dy;
+	float marginLeft, marginTop, dx, dy;
 };
